+++ test/fixtures/framework/features/support/env.rb
@@ -1,18 +1,3 @@
-# Verifies that an environment variable is set to a given value.
-#
-# @step_input key [String] The environment variable
-# @step_input value [String] The intended value of the environment variable
-Then('the Runner.environment entry for {string} equals {string}') do |key, expected_value|
-  assert_equal(expected_value, Maze::Runner.environment[key])
-end
-
-# Verifies that an environment variable is not set.
-#
-# @step_input key [String] The environment variable
-Then('the Runner.environment entry for {string} is null') do |key|
-  assert_nil(Maze::Runner.environment[key])
-end
-
 Maze.hooks.before_all do
   Maze::Runner.environment['AFTER_CONFIG'] = 'FIRST_SCENARIO_ONLY'
 end