+++ test/server_test.rb
@@ -3,6 +3,10 @@
 require 'cucumber'
 require 'test_helper'
 require 'webrick'
+require_relative '../lib/maze/servlets/base_servlet'
+require_relative '../lib/maze/servlets/command_servlet'
+require_relative '../lib/maze/servlets/log_servlet'
+require_relative '../lib/maze/servlets/servlet'
 require_relative '../lib/maze/server'
 
 # noinspection RubyNilAnalysis
@@ -35,6 +39,7 @@ module Maze
       mock_http_server.expects(:mount).with('/uploads', any_parameters).once
       mock_http_server.expects(:mount).with('/sourcemap', any_parameters).once
       mock_http_server.expects(:mount).with('/react-native-source-map', any_parameters).once
+      mock_http_server.expects(:mount).with('/command', any_parameters).once
       mock_http_server.expects(:mount).with('/logs', any_parameters).once
 
       mock_http_server.expects(:start)
@@ -79,6 +84,7 @@ module Maze
       mock_http_server.expects(:mount).with('/sourcemap', any_parameters).once
       mock_http_server.expects(:mount).with('/react-native-source-map', any_parameters).once
       mock_http_server.expects(:mount).with('/logs', any_parameters).once
+      mock_http_server.expects(:mount).with('/command', any_parameters).once
       mock_http_server.expects(:start)
       mock_http_server.expects(:shutdown)
       WEBrick::HTTPServer.expects(:new).with(Port: PORT,